refactored model_args and model protocol location
